+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+* Fix proxying to `https` target `hotel add https://example.com`
+* Log proxy errors
+
 ## 0.5.13
 
 * Fix `hotel add` CLI bug